Se modifico la creación del tópico para que sea realizada por el usuario autenticado por token y se modificó para la modificacion y eliminacion de un tópico que solo pueda hacerlo el usuario que lo haya creado

+++ foro-hub/src/main/java/com/github/matialvarez7/foro_hub/domain/topico/TopicoService.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -24,15 +25,8 @@ public class TopicoService {
     private UsuarioRepository usuarioRepository;
 
 
-    public RegistroTopicoResponse registrarTopico(@Valid RegistroTopicoRequest datos) {
+    public RegistroTopicoResponse registrarTopico(@Valid RegistroTopicoRequest datos, Usuario usuarioAutenticado) {
 
-        Usuario autor;
-        Optional<Usuario> usuario = usuarioRepository.findById(datos.idUsuario());
-        if (usuario.isPresent()){
-            autor = usuario.get();
-        } else {
-            throw new ValidacionException("El id de usuario no existe");
-        }
         Curso curso = cursoRepository.findByNombreEquals(datos.nombreCurso());
         if (curso == null) {
             throw new ValidacionException("El curso no existe");
@@ -42,7 +36,7 @@ public class TopicoService {
             throw new ValidacionException("Ya existe un tópico similar en el foro");
         }
 
-        nuevoTopico = new Topico(datos, autor, curso);
+        nuevoTopico = new Topico(datos, usuarioAutenticado, curso);
         topicoRepository.save(nuevoTopico);
         return new RegistroTopicoResponse(nuevoTopico);
     }
@@ -60,24 +54,33 @@ public class TopicoService {
         }
     }
 
-    public ModificacionTopicoResponse actualizarTopico(Long id, ModificacionTopicoRequest datosModificacion) {
+    public ModificacionTopicoResponse actualizarTopico(Long id, ModificacionTopicoRequest datosModificacion, Usuario usuarioAutenticado) {
         if (datosModificacion.titulo() == null && datosModificacion.mensaje() == null) {
             throw new ValidacionException("No se enviaron datos para realizar modificación");
         }
         Optional<Topico> topicoBuscado = topicoRepository.findById(id);
         if(topicoBuscado.isPresent()){
             Topico topico = topicoBuscado.get();
-            topico.actualizarTopico(datosModificacion);
-            return new ModificacionTopicoResponse(topico);
+            if (Objects.equals(topico.getAutor().getId(), usuarioAutenticado.getId())){
+                topico.actualizarTopico(datosModificacion);
+                return new ModificacionTopicoResponse(topico);
+            } else {
+                throw new ValidacionException("No tiene permisos para modificar este tópico.");
+            }
         } else {
             throw new ValidacionException("No existe el topico que desea modificar");
         }
     }
 
-    public void eliminarTopico(Long id) {
+    public void eliminarTopico(Long id, Usuario usuarioAutenticado) {
         Optional<Topico> topicoBuscado = topicoRepository.findById(id);
         if (topicoBuscado.isPresent()) {
-            topicoRepository.deleteById(id);
+            if(Objects.equals(topicoBuscado.get().getAutor().getId(), usuarioAutenticado.getId())) {
+                topicoRepository.deleteById(id);
+            }
+            else {
+                throw new ValidacionException("No tiene permisos para eliminar este tópico.");
+            }
         } else {
             throw new ValidacionException("No existe el topico que desea eliminar");
         }
